# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Student Management API",
        "description": "API для управления студентами, группами и лекциями",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/lectures": {
            "get": {
                "tags": [
                    "Lectures"
                ],
                "summary": "Получить список всех лекций",
                "operationId": "cfa73133376ac38dad9b363e6e82246d",
                "responses": {
                    "200": {
                        "description": "Успешное получение списка лекций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "topic": {
                                                "type": "string",
                                                "example": "Введение в программирование"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Основные понятия программирования"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Lectures"
                ],
                "summary": "Создать новую лекцию",
                "operationId": "c93d4feb5eac2deb7ca967f21fee3a2f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "topic",
                                    "description"
                                ],
                                "properties": {
                                    "topic": {
                                        "type": "string",
                                        "example": "Введение в программирование"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Основные понятия программирования"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Лекция успешно создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "topic": {
                                            "type": "string",
                                            "example": "Введение в программирование"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Основные понятия программирования"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "topic": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The topic has already been taken."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/lectures/group/{groupId}": {
            "get": {
                "tags": [
                    "Lectures"
                ],
                "summary": "Получить лекции по группе",
                "operationId": "9740ff6671c25f7884348e51a4b9a756",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "ID группы",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список лекций группы",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "topic": {
                                                "type": "string",
                                                "example": "Введение в программирование"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Основные понятия программирования"
                                            },
                                            "order": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/lectures/{lectureId}/attend/{studentId}": {
            "post": {
                "tags": [
                    "Lectures"
                ],
                "summary": "Отметить посещение лекции студентом",
                "operationId": "8ca75a85dade58553f5aa9d0a0193889",
                "parameters": [
                    {
                        "name": "lectureId",
                        "in": "path",
                        "description": "ID лекции",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "studentId",
                        "in": "path",
                        "description": "ID студента",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Посещение отмечено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lecture attendance marked"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Лекция или студент не найдены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Lecture or student not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}